use renamed example json files
